test(billing): align coverage with removed batch props
